requirements.txt
django==3.2.7
environs==9.5.*
Pillow==9.1.*
django-admin-sortable2==1.0
django-tinymce==3.4.0
